feat: automatic social account group creation
When enabling group sync with a third party auth provider, groups from e.g. the
OIDC response are added to the user if they're already present in paperless.
This presents a hassle in complex SSO environments: you don't just need to
create groups and add users in your authentication provider, you also need to
go to paperless and add them there.

This PR adds a config flag that allows automated creation of groups based on
OIDC responses. Groups are not automatically vacuumed if they're empty.